commit final site version EXAMEN

+++ app/Http/Controllers/FAQController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\FAQ;
+use Auth;
 
 class FAQController extends Controller
 {
@@ -20,7 +21,7 @@ class FAQController extends Controller
         return view('Pages.FAQ', ['faqs' => $faqs]);
     }
 
-        // ___________________________________________________________________________
+    // ___________________________________________________________________________
     /**
      * Show the form for editing the specified resource.
      *
@@ -29,7 +30,10 @@ class FAQController extends Controller
      */
     public function edit(FAQ $faq)
     {
-        return view('Pages.Admin.FAQModif',['faq'=>$faq]);
+        if (Auth::user()->isAdmin()) {
+            return view('Pages.Admin.FAQModif', ['faq' => $faq]);
+        }
+        abort(403);
     }
 
     // ___________________________________________________________________________
@@ -42,19 +46,22 @@ class FAQController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate([
-            'question' => 'required',
-            'reponse' => 'required',
-        ]);
+        if (Auth::user()->isAdmin()) {
+            $request->validate([
+                'question' => 'required',
+                'reponse' => 'required',
+            ]);
 
-        FAQ::create([
-            'question' => $request->question,
-            'reponse' => $request->reponse
-        ]);
-        return redirect()->route('admin.index')->with('status', 'La nouvelle Question/Réponse a bien été ajoutée');
+            FAQ::create([
+                'question' => $request->question,
+                'reponse' => $request->reponse
+            ]);
+            return redirect()->route('admin.index')->with('status', 'La nouvelle Question/Réponse a bien été ajoutée');
+        }
+        abort(403);
     }
 
-        // ___________________________________________________________________________
+    // ___________________________________________________________________________
     /**
      * Update the specified resource in storage.
      *
@@ -64,22 +71,23 @@ class FAQController extends Controller
      */
     public function update(Request $request, FAQ $faq)
     {
+        if (Auth::user()->isAdmin()) {
+            $request->validate([
+                'question' => 'required',
+                'reponse' => 'required',
+            ]);
 
-        $request->validate([
-            'question' => 'required',
-            'reponse' => 'required',
-        ]);
+            $faq->update([
+                'question' => $request->input('question'),
+                'reponse' => $request->input('reponse'),
+            ]);
 
-        $faq->update([
-            'question' => $request->input('question'),
-            'reponse' => $request->input('reponse'),
-        ]);
+            $faq->save();
 
-        $faq->save();
-
-        return redirect()->route('admin.index')->with('status', 'La mise à jour a bien été effectuée');
+            return redirect()->route('admin.index')->with('status', 'La mise à jour a bien été effectuée');
+        }
+        abort(403);
     }
-
 
     // ___________________________________________________________________________
     /**
@@ -90,8 +98,11 @@ class FAQController extends Controller
      */
     public function destroy($id)
     {
-        $faq = FAQ::find($id);
-        $faq->delete();
-        return redirect()->route('admin.index')->with('status', 'La question/réponse a bien été supprimée');
+        if (Auth::user()->isAdmin()) {
+            $faq = FAQ::find($id);
+            $faq->delete();
+            return redirect()->route('admin.index')->with('status', 'La question/réponse a bien été supprimée');
+        }
+        abort(403);
     }
 }
